creates deprecations for the links with icons block
