+++ private/src/scripts/editor/blocks/links-with-icons/index.jsx
@@ -1,13 +1,9 @@
 /**
- * Third-party
- */
-import classnames from 'classnames';
-
-/**
  * Module-specific
  */
 import DisplayComponent from './DisplayComponent.jsx';
 import './inner-block.jsx';
+import deprecated from './deprecated.jsx';
 
 /**
  * WordPress
@@ -49,47 +45,7 @@ registerBlockType('amnesty-core/repeatable-block', {
       default: 'none',
     },
   },
-  deprecated: [
-    {
-      attributes: {
-        backgroundColor: {
-          type: 'string',
-        },
-        orientation: {
-          type: 'string',
-          default: 'horizontal',
-        },
-        quantity: {
-          type: 'number',
-          default: 2,
-        },
-        hideLines: {
-          type: 'boolean',
-          default: false,
-        },
-      },
-      save({ attributes, className }) {
-        const { quantity, orientation = 'horizontal', backgroundColor, hideLines } = attributes;
-        const classes = classnames(
-          'linksWithIcons-group',
-          `is-${orientation}`,
-          `has-${quantity}-items`,
-          {
-            'has-background': !!backgroundColor,
-            [`has-${backgroundColor}-background-color`]: !!backgroundColor,
-            'has-no-lines': !!hideLines,
-            className: !!className,
-          },
-        );
-
-        return (
-          <div className={classes}>
-            <InnerBlocks.Content />
-          </div>
-        );
-      },
-    },
-  ],
+  deprecated,
   edit: DisplayComponent,
 
   save: () => <InnerBlocks.Content />,
